fix: remove duplicate route and correct param usage
