+++ app/(user)/assessments/[assessmentId]/control-workspace/[controlId]/page.tsx
@@ -5,26 +5,38 @@ import { useParams } from "next/navigation";
 import ControlWorkspace from "@/components/user/control-workspace/ControlWorkspace";
 import { getControl } from "@/services/control.services";
 
+interface ControlWorkspaceData {
+  id: string;
+  framework: string;
+  title: string;
+  description: string;
+  severity: string;
+  status: string;
+  weight: number;
+  assessmentId: string;
+}
 
 export default function Page() {
   const params = useParams();
 
-  const assessmentId = params?.assessmentId as string;
+  const assessmentId = params?.assessmentId as string; // ✅ correct
   const controlId = params?.controlId as string;
 
-  const [control, setControl] = useState<any>(null);
+  const [control, setControl] = useState<ControlWorkspaceData | null>(null);
   const [loading, setLoading] = useState(true);
 
-  // 🧠 Fetch control data
   useEffect(() => {
-    if (!assessmentId || !controlId) return;
+    if (!assessmentId || !controlId) {
+      return;
+    }
 
     const fetchData = async () => {
       try {
         const data = await getControl(controlId);
+
         setControl({
           ...data,
-          assessmentId, // attach route param if needed
+          assessmentId,
         });
       } catch (err) {
         console.error("Failed to load control:", err);
@@ -36,21 +48,17 @@ export default function Page() {
     fetchData();
   }, [assessmentId, controlId]);
 
-  // 🛑 Invalid route
   if (!assessmentId || !controlId) {
     return <div>Invalid route</div>;
   }
 
-  // ⏳ Loading state
   if (loading) {
     return <div className="p-6 text-sm text-muted-foreground">Loading control...</div>;
   }
 
-  // ❌ Safety fallback
   if (!control) {
     return <div className="p-6 text-sm text-destructive">Failed to load control</div>;
   }
 
   return <ControlWorkspace control={control} />;
 }
-
